## 🎯 Overview

Implementación completa del flujo de pagos para Mini App Web con
detección automática de pagos y historial de transacciones.

## 🚀 Cambios Principales

### Backend APIs
- ✅ `GET /api/v1/miniapp/api/payment-status/{transaction_id}` - Endpoint
para polling de estado de pagos
- ✅ `GET /api/v1/miniapp/transactions` - Historial de transacciones con
paginación y filtros
- ✅ Repository methods actualizados para DataPackage, CryptoOrder y
Subscription

### Bot Integration
- ✅ Detección de pagos Mini App via payload prefix `miniapp_*`
- ✅ Nuevo handler `_handle_miniapp_successful_payment()` para packages,
slots y subscriptions
- ✅ Mensaje de confirmación con botón "🚀 Abrir Mini App"
- ✅ Backward compatibility con payloads existentes

### Frontend Mini App
- ✅ Auto-polling cada 3 segundos para detectar pago completado
- ✅ Waiting screen con información de transacción
- ✅ Timeout automático a los 5 minutos
- ✅ Haptic feedback para success/error
- ✅ Auto-redirect a profile después del pago
- ✅ Página dedicada de transacciones con filtros y paginación

## 📦 Archivos Modificados

### Backend
- `infrastructure/api/routes/miniapp_payments.py` - Payment status +
transactions endpoints
- `infrastructure/persistence/postgresql/*.py` - Repository methods
- `domain/interfaces/*.py` - Repository interfaces

### Bot
- `telegram_bot/features/buy_gb/handlers_confirmation.py` - Mini App
payment detection
- `tests/telegram_bot/features/buy_gb/test_handlers_confirmation.py` -
Tests

### Frontend
- `miniapp/templates/purchase.html` - Auto-polling JavaScript
- `miniapp/templates/transactions.html` - Nueva página de transacciones
- `miniapp/templates/base.html` - Navigation link
- `miniapp/templates/profile.html` - Link a historial completo

## 🧪 Test Coverage

| Suite | Tests | Estado |
|-------|-------|--------|
| Mini App Payment Endpoints | 24 | ✅ Passing |
| Mini App Transaction History | 7 | ✅ Passing |
| Mini App Entry/Profile/Privacy | 34 | ✅ Passing |
| Bot Payment Handlers | 4 | ✅ Passing |
| **TOTAL** | **69** | ✅ **All Passing** |

## 🎨 Flujo de Usuario

```
1. Usuario selecciona producto en Mini App
2. Elige método de pago (Stars/Crypto)
3. Paga en Telegram (Stars) o vía QR (Crypto)
4. Bot detecta pago exitoso → Registra en DB
5. Mini App polls → Detecta "completed"
6. Muestra success → Redirect a profile
7. Usuario ve transacción en historial
```

## 📝 Commits

- feat: Migrate Mini App API routes to /api/v1/miniapp
- refactor: Clean up old Mini App routes and finalize migration
- docs: Add Mini App payment flow implementation plan
- feat: Add payment status polling endpoint for Mini App
- feat: Add transaction history endpoint with pagination
- feat: Detect and handle Mini App payments in Bot
- feat: Add auto-polling for payment status in Mini App
- feat: Create dedicated transactions history page

## ✅ Checklist

- [x] Código implementado y testeado
- [x] Tests passing (69/69)
- [x] Documentación actualizada
- [x] Backward compatibility mantenida
- [x] No breaking changes
- [x] Ready for review

## 🔗 Related Issues

- Closes: Mini App Payment Flow Implementation
- Related: #281 (Subscription Plans Implementation)

+++ infrastructure/persistence/postgresql/models/crypto_transaction.py
@@ -2,9 +2,9 @@ import uuid
 from datetime import datetime, timezone
 from typing import Any, Dict, Optional
 
-from sqlalchemy import JSON, Column, DateTime
+from sqlalchemy import JSON, BigInteger, DateTime
 from sqlalchemy import Enum as SQLEnum
-from sqlalchemy import Float, ForeignKey, Integer, String
+from sqlalchemy import Float, ForeignKey, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -21,7 +21,7 @@ class CryptoTransactionModel(Base):
 
     id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     user_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("users.telegram_id"), nullable=False, index=True
+        BigInteger, ForeignKey("users.telegram_id"), nullable=False, index=True
     )
     wallet_address: Mapped[str] = mapped_column(String(42), nullable=False)
     amount: Mapped[float] = mapped_column(Float, nullable=False)
